+++ faststream/nats/broker/__init__.py
@@ -1,0 +1,3 @@
+from faststream.nats.broker.broker import NatsBroker
+
+__all__ = ("NatsBroker",)
