+++ faststream/nats/router/__init__.py
@@ -1,0 +1,6 @@
+from faststream.nats.router.router import NatsRoute, NatsRouter
+
+__all__ = (
+    "NatsRoute",
+    "NatsRouter",
+)
