+++ faststream/nats/router/router.pyi
